Fix line endings & library import
