+++ examples/doc/basic.html
