+++ examples/doc/block-inline.html
