+++ package.json
@@ -2,7 +2,7 @@
   "name": "html2slack",
   "version": "1.0.1",
   "description": "A js library to convert a subset of HTML to JSON formatted Slack attachments",
-  "main": "src/index.js",
+  "main": "src/lib",
   "directories": {
     "test": "test"
   },
@@ -29,6 +29,9 @@
     "romanize": "^1.1.1"
   },
   "devDependencies": {
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.11.0",
     "jest": "^22.0.4"
   }
 }
